+++ src/domain/sensor_domain/models/sensor.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <memory_resource>
 #include <string>
 
@@ -16,7 +17,7 @@ struct Sensor {
     using allocator_type = std::pmr::polymorphic_allocator<>;
 
     std::pmr::string id;
-    size_t id_hash = 0;
+    uint32_t id_hash = 0;
     SensorMetadata metadata;
     SensorConfiguration configuration;
