+++ src/domain/sensor_domain/processors/collect_isr_reading_processor.h
@@ -16,7 +16,7 @@ private:
 public:
     explicit CollectIsrReadingProcessor(std::shared_ptr<SensorReadingsFrame> sensor_readings_frame);
 
-    void ProcessReading(const size_t sensor_id_hash) override;
+    void ProcessReading(const uint32_t sensor_id_hash) override;
 };
 
 } // namespace eerie_leap::domain::sensor_domain::processors
