+++ src/domain/sensor_domain/processors/script_processor.cpp
@@ -7,7 +7,7 @@ using namespace eerie_leap::domain::sensor_domain::models;
 ScriptProcessor::ScriptProcessor(const std::string& function_name, std::shared_ptr<SensorReadingsFrame> sensor_readings_frame)
     : function_name_(function_name), sensor_readings_frame_(std::move(sensor_readings_frame)) {}
 
-void ScriptProcessor::ProcessReading(const size_t sensor_id_hash) {
+void ScriptProcessor::ProcessReading(const uint32_t sensor_id_hash) {
     auto reading_optioanl = sensor_readings_frame_->TryGetReading(sensor_id_hash);
     if(!reading_optioanl)
         return;
